refactor: 테스트용 Mock 클래스 네이밍 Stub 으로 변경
