+++ src/test/java/library/book/presentation/utils/FunctionMangerTest.java
@@ -6,12 +6,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
 
-import library.book.mock.MockFunctionExecutor;
+import library.book.stub.StubFunctionExecutor;
 
 @DisplayName("[FunctionManger Test] - Presentation")
 public class FunctionMangerTest {
 
-	private final FunctionExecutor functionExecutor = new MockFunctionExecutor(null, null);
+	private final FunctionExecutor functionExecutor = new StubFunctionExecutor(null, null);
 
 	@Test
 	@DisplayName("[ONE 테스트]")
